Port PkiCertProcess to libtoolchain.
